requirements.txt
flask==2.3.2
python-dotenv==1.0.0
flask-sqlalchemy==3.0.3
requests==2.31.0